# ApiServices/docs/openapi.json
{"openapi":"3.0.2","info":{"title":"Secure AI Labs API Services","description":"All the private and public APIs for the Secure AI Labs","version":"0.1.0"},"paths":{"/login":{"post":{"summary":"Login For Access Token","description":"User login with email and password","operationId":"login_for_access_token_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login_Out"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/organizations":{"get":{"summary":"Get All Organizations","description":"Get list of all the organizations","operationId":"get_all_organizations_organizations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Organizations Organizations Get","type":"array","items":{"$ref":"#/components/schemas/GetOrganizations_Out"}}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"summary":"Register Organization","description":"Register new organization and the admin user","operationId":"register_organization_organizations_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterOrganization_In"}}},"required":true},"responses":{"200":{"description":"Organization Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterOrganization_Out"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/organizations/{organization_id}":{"get":{"summary":"Get Organization","description":"Get the information about a organization","operationId":"get_organization_organizations__organization_id__get","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetOrganizations_Out"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"summary":"Update Organization","description":"Update organization","operationId":"update_organization_organizations__organization_id__put","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrganization_In"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"summary":"Delete Organization","description":"Delete organization","operationId":"delete_organization_organizations__organization_id__delete","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/organizations/{organization_id}/users":{"get":{"summary":"Get Users","description":"Get all users in the organization","operationId":"get_users_organizations__organization_id__users_get","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Users Organizations  Organization Id  Users Get","type":"array","items":{"$ref":"#/components/schemas/GetUsers_Out"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"summary":"Register User","description":"Add new user to organization","operationId":"register_user_organizations__organization_id__users_post","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUser_In"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUser_Out"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/organizations/{organization_id}/users/{user_id}":{"get":{"summary":"Get User","description":"Get information about a user","operationId":"get_user_organizations__organization_id__users__user_id__get","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"},{"required":true,"schema":{"title":"User Id","type":"string","format":"uuid"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsers_Out"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"summary":"Update User","description":"Update user information","operationId":"update_user_organizations__organization_id__users__user_id__put","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"},{"required":true,"schema":{"title":"User Id","type":"string","format":"uuid"},"name":"user_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUser_In"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"summary":"Delete User","description":"Delete user","operationId":"delete_user_organizations__organization_id__users__user_id__delete","parameters":[{"required":true,"schema":{"title":"Organization Id","type":"string","format":"uuid"},"name":"organization_id","in":"path"},{"required":true,"schema":{"title":"User Id","type":"string","format":"uuid"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_login_post":{"title":"Body_login_for_access_token_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"GetOrganizations_Out":{"title":"GetOrganizations_Out","required":["name","address","primaryContactName","primaryContactTitle","primaryContactEmail","primaryContactPhone","secondaryContactName","secondaryContactTitle","secondaryContactEmail","secondaryContactPhone","_id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"address":{"title":"Address","type":"string"},"primaryContactName":{"title":"Primarycontactname","type":"string"},"primaryContactTitle":{"title":"Primarycontacttitle","type":"string"},"primaryContactEmail":{"title":"Primarycontactemail","type":"string","format":"email"},"primaryContactPhone":{"title":"Primarycontactphone","type":"string"},"secondaryContactName":{"title":"Secondarycontactname","type":"string"},"secondaryContactTitle":{"title":"Secondarycontacttitle","type":"string"},"secondaryContactEmail":{"title":"Secondarycontactemail","type":"string","format":"email"},"secondaryContactPhone":{"title":"Secondarycontactphone","type":"string"},"_id":{"title":" Id","type":"string","format":"uuid"}}},"GetUsers_Out":{"title":"GetUsers_Out","required":["username","email","title","phone","role","_id"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"title":{"title":"Title","type":"string"},"phone":{"title":"Phone","type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"_id":{"title":" Id","type":"string","format":"uuid"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Login_Out":{"title":"Login_Out","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"RegisterOrganization_In":{"title":"RegisterOrganization_In","required":["name","address","primaryContactName","primaryContactTitle","primaryContactEmail","primaryContactPhone","secondaryContactName","secondaryContactTitle","secondaryContactEmail","secondaryContactPhone","primaryContactPassword"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"address":{"title":"Address","type":"string"},"primaryContactName":{"title":"Primarycontactname","type":"string"},"primaryContactTitle":{"title":"Primarycontacttitle","type":"string"},"primaryContactEmail":{"title":"Primarycontactemail","type":"string","format":"email"},"primaryContactPhone":{"title":"Primarycontactphone","type":"string"},"secondaryContactName":{"title":"Secondarycontactname","type":"string"},"secondaryContactTitle":{"title":"Secondarycontacttitle","type":"string"},"secondaryContactEmail":{"title":"Secondarycontactemail","type":"string","format":"email"},"secondaryContactPhone":{"title":"Secondarycontactphone","type":"string"},"primaryContactPassword":{"title":"Primarycontactpassword","type":"string"}}},"RegisterOrganization_Out":{"title":"RegisterOrganization_Out","required":["_id"],"type":"object","properties":{"_id":{"title":" Id","type":"string","format":"uuid"}}},"RegisterUser_In":{"title":"RegisterUser_In","required":["username","email","title","phone","role","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"title":{"title":"Title","type":"string"},"phone":{"title":"Phone","type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"password":{"title":"Password","type":"string"}}},"RegisterUser_Out":{"title":"RegisterUser_Out","required":["username","email","title","phone","role","_id","accountState"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"title":{"title":"Title","type":"string"},"phone":{"title":"Phone","type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"_id":{"title":" Id","type":"string","format":"uuid"},"accountState":{"$ref":"#/components/schemas/UserAccountState"}}},"UpdateOrganization_In":{"title":"UpdateOrganization_In","required":["primaryContactName","primaryContactTitle","primaryContactEmail","primaryContactPhone","secondaryContactName","secondaryContactTitle","secondaryContactEmail","secondaryContactPhone"],"type":"object","properties":{"primaryContactName":{"title":"Primarycontactname","type":"string"},"primaryContactTitle":{"title":"Primarycontacttitle","type":"string"},"primaryContactEmail":{"title":"Primarycontactemail","type":"string","format":"email"},"primaryContactPhone":{"title":"Primarycontactphone","type":"string"},"secondaryContactName":{"title":"Secondarycontactname","type":"string"},"secondaryContactTitle":{"title":"Secondarycontacttitle","type":"string"},"secondaryContactEmail":{"title":"Secondarycontactemail","type":"string","format":"email"},"secondaryContactPhone":{"title":"Secondarycontactphone","type":"string"}}},"UpdateUser_In":{"title":"UpdateUser_In","required":["phone","role","accountState"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"phone":{"title":"Phone","type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"accountState":{"$ref":"#/components/schemas/UserAccountState"}}},"UserAccountState":{"title":"UserAccountState","enum":["Active","Inactive","Locked"],"description":"An enumeration."},"UserRole":{"title":"UserRole","enum":["Admin","Auditor","User","DigitalContractAdmin","DatasetAdmin","SailAdmin"],"description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}